chore(voucher-management): add check null prevention

+++ DAL/DAL_Voucher.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -12,24 +13,41 @@ namespace DAL
         ShopDatabaseEntities db = new ShopDatabaseEntities();
         public void Them(tblVoucher Voucher)
         {
-            db.tblVouchers.Add(Voucher);
-            db.SaveChanges();
+            if (Voucher != null)
+            {
+                db.tblVouchers.Add(Voucher);
+                db.SaveChanges();
+            }
         }
         public void Sua(tblVoucher OldVoucher, tblVoucher NewVoucher)
         {
-            tblVoucher Voucher = db.tblVouchers.ToList().Find(vc => Equals(vc.MaV, OldVoucher.MaV));
-            Voucher.TenV = NewVoucher.TenV;
-            Voucher.MoTa = NewVoucher.MoTa;
-            Voucher.GiaTri = NewVoucher.GiaTri;
-            Voucher.DonVi = NewVoucher.DonVi;
-            Voucher.GTToiThieu = NewVoucher.GTToiThieu;
-            Voucher.GTToiDa = NewVoucher.GTToiDa;
+            if (NewVoucher is null)
+            {
+                return;
+            }
+            if (OldVoucher is null)
+            {
+                Them(NewVoucher);
+            }
+            else
+            {
+                tblVoucher Voucher = db.tblVouchers.ToList().Find(vc => Equals(vc.MaV, OldVoucher.MaV));
+                Voucher.TenV = NewVoucher.TenV;
+                Voucher.MoTa = NewVoucher.MoTa;
+                Voucher.GiaTri = NewVoucher.GiaTri;
+                Voucher.DonVi = NewVoucher.DonVi;
+                Voucher.GTToiThieu = NewVoucher.GTToiThieu;
+                Voucher.GTToiDa = NewVoucher.GTToiDa;
+            }
             db.SaveChanges();
         }
         public void Xoa(tblVoucher Voucher)
         {
-            db.tblVouchers.Remove(Voucher);
-            db.SaveChanges();
+            if (Voucher != null)
+            {
+                db.tblVouchers.Remove(Voucher);
+                db.SaveChanges();
+            }
         }
         public List<tblVoucher> DanhSachVoucher() => db.tblVouchers.ToList();
     }
